Update launcher version and fixes

+++ src/update-handling/LauncherStrategy.ts
@@ -329,10 +329,18 @@ export abstract class LauncherStrategy {
     }
 
     public async setCustomFont(value: boolean) {
+        if (!this.w3PathIsValid) return;
         if (value) {
-            fs.rmdirSync(`${this.w3Path}/fonts`, {recursive: true});
-        } else {
             this.downloadAndWriteFile("fonts", this.w3Path);
+            if (this.isTest) {
+                this.downloadAndWriteFile("fonts", this.w3Path.replace('retail', 'ptr'))
+            }
+        } else {
+            fs.rmdirSync(`${this.w3Path}/fonts`, {recursive: true});
+            if (this.isTest) {
+                fs.rmdirSync(`${this.w3Path.replace('retail', 'ptr')}/fonts`,
+                {recursive: true});
+            }
         }
     }
